// src/lib/types/running.ts
export interface RunningTraining {
    distance: number;
    moving_time: number;
    date: Date;
}